{{cookiecutter.project_slug}}/requirements.txt
# Python

amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
billiard==4.1.0
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect==1.2.0
cssutils==2.7.1
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
emails==0.6
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
premailer==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==2.3.0
pydantic-settings==2.0.2
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.1
vine==5.0.0
wcwidth==0.2.6
boto3==1.28.74
slowapi==0.1.8

# Linters
ruff==0.1.11
pre-commit==3.5.0
conventional-pre-commit==3.2.0

# Database

alembic==1.11.2
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.2

# Tests

pytest==7.4.0
pytest-cov==4.1.0
pytest-sugar==0.9.7
pytest-icdiff==0.9
pytest-xdist==3.6.1
coverage==7.3.0

# Redis

redis==5.0.1
